% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{append_latest_prices}
\alias{append_latest_prices}
\title{Append most recent prices to existing files}
\usage{
append_latest_prices(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Append prices for all tickers in the specified path. The existing
csv files are used.
}
