% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{append_latest_prices_db}
\alias{append_latest_prices_db}
\title{Append most recent prices to existing files}
\usage{
append_latest_prices_db(db_path, tickers)
}
\arguments{
\item{db_path}{A single character string. Path where data are stored.}

\item{tickers}{A vector of character strings containing tickers.}
}
\description{
Append prices for all tickers specified as an argument. The existing
csv files are used to append prices.
}
