% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Format column names of a data frame}
\usage{
clean_column_names(df)
}
\arguments{
\item{df}{A data frame with unformatted column names.}
}
\value{
A data frame with formatted column names.
}
\description{
Format column names of a data frame
}
