% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_investment_names}
\alias{clean_investment_names}
\title{Clean names of investments}
\usage{
clean_investment_names(investment_names)
}
\arguments{
\item{investment_names}{A single character string.}
}
\value{
A string with cleaned names.
}
\description{
Clean names of investments
}
