% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_annualized_returns}
\alias{compute_annualized_returns}
\title{Compute annualized returns based on daily returns for matrix or xts}
\usage{
compute_annualized_returns(R, scale = 252)
}
\arguments{
\item{R}{An xts or matrix with returns}

\item{scale}{A numeric for the number of periods in a year (daily = 252 is
default, monthly = 12, quarterly = 4, yearly = 1)}
}
\value{
A data frame with annualized returns for all investments
}
\description{
Compute annualized returns based on daily returns for matrix or xts
}
