% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_splits.R
\name{correct_quantity_panels_for_stock_splits}
\alias{correct_quantity_panels_for_stock_splits}
\title{Get quantity panel for ticker corrected for stock splits}
\usage{
correct_quantity_panels_for_stock_splits(
         df_panel, ticker, df_transactions_with_tickers, user_path)
}
\arguments{
\item{df_panel}{A data frame which contains a quantity panel for a given ticker.}

\item{ticker}{A single character string containing the ticker symbol.}

\item{df_transactions_with_tickers}{A data frame containing transactions
and ticker.}

\item{user_path}{A single character string containing the directory of the user.}
}
\value{
A data frame which contains a quantity panel corrected for stock splits.
}
\description{
Get quantity panel for ticker corrected for stock splits
}
