% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfoliotracker_dir.R
\name{create_main_dir}
\alias{create_main_dir}
\title{Create main directory of the Portfolio Tracker}
\usage{
create_main_dir(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\value{
Returns a list with path names for data storage
}
\description{
Create the main directory of the Portfolio Tracker with all required folders
}
