% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfoliotracker_files.R
\name{create_main_files}
\alias{create_main_files}
\title{Create main files of the Portfolio Tracker}
\usage{
create_main_files(path)
}
\arguments{
\item{path}{A single character string. Path where files are created.}
}
\description{
Create the main files of the Portfolio Tracker. Create files for cookies,
user database,
}
