% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{create_quantity_panel}
\alias{create_quantity_panel}
\title{Create quantity panel for ticker}
\usage{
create_quantity_panel(ticker, df.transaction.history, path.quantitypanel)
}
\arguments{
\item{ticker}{A single character string containing the ticker symbol.}

\item{df.transaction.history}{A data.frame containing the full history of transactions.}

\item{path.quantitypanel}{A single character string containing the folder of quantity panels.}
}
\value{
\emph{create_quantity_panel} returns
}
\description{
Create quantity panel for ticker
}
