% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfoliotracker_dir.R
\name{create_user_dir}
\alias{create_user_dir}
\title{Create directory of the user}
\usage{
create_user_dir(path, portfolio_name = "portfolio_1")
}
\arguments{
\item{path}{A single character string. Path where user data are stored.}

\item{portfolio_name}{A single character string indicating the name of the
portfolio. Currently the default is \emph{portfolio_1}.}
}
\value{
Returns a list with path names for user-specific data storage
}
\description{
Create directory of the user
}
