% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formatting.R
\name{format_current_portfolio}
\alias{format_current_portfolio}
\title{Format current portfolio}
\usage{
format_current_portfolio(df_investments)
}
\arguments{
\item{df_investments}{A data frame containing current portfolio.}
}
\value{
A data frame containing the current portfolio with formatted
names and numbers.
}
\description{
Format current portfolio
}
