% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formatting.R
\name{format_previous_investments}
\alias{format_previous_investments}
\title{Format previous investment portfolio}
\usage{
format_previous_investments(df_investments)
}
\arguments{
\item{df_investments}{A data frame containing previous investments}
}
\value{
A data frame containing previous investments with formatted
names and numbers.
}
\description{
Format previous investment portfolio
}
