% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formatting.R
\name{format_rebalanced_portfolio}
\alias{format_rebalanced_portfolio}
\title{Format rebalanced portfolio}
\usage{
format_rebalanced_portfolio(df_investments)
}
\arguments{
\item{df_investments}{A data frame containing a rebalanced portfolio.}
}
\value{
A data frame containing the rebalanced portfolio with
formatted numbers.
}
\description{
Format rebalanced portfolio
}
