% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formatting.R
\name{format_return_tables}
\alias{format_return_tables}
\title{Format table with returns}
\usage{
format_return_tables(path, df_returns)
}
\arguments{
\item{path}{A single character string. Directory of the data.}

\item{df_returns}{A data frame with table structure of returns,
i.e., ticker and time periods.}
}
\value{
A data frame containing formatted table with returns.
}
\description{
Format table with returns
}
