% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formatting.R
\name{format_stock_splits}
\alias{format_stock_splits}
\title{Format current or previous stock splits}
\usage{
format_stock_splits(df_stock_splits)
}
\arguments{
\item{df_stock_splits}{A data frame containing current or previous
stock splits}
}
\value{
A data frame containing the current or previous stock splits
with formatted names and numbers.
}
\description{
Format current or previous stock splits
}
