% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returns.R
\name{get_annual_returns}
\alias{get_annual_returns}
\title{Get annual returns}
\usage{
get_annual_returns(path)
}
\arguments{
\item{path}{A single character string. Directory of your data.}
}
\value{
A data frame containing current portfolio.
}
\description{
Get annual returns
}
