% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_annualized_return}
\alias{get_annualized_return}
\title{Get annualized return based on daily returns for vector}
\usage{
get_annualized_return(R, scale = 252)
}
\arguments{
\item{R}{An xts or vector with returns}

\item{scale}{A numeric for the number of periods in a year (daily = 252 is
default, monthly = 12, quarterly = 4, yearly = 1)}
}
\value{
A vector with annualized returns
}
\description{
Get annualized return based on daily returns for vector
}
