% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_price_date_range}
\alias{get_available_price_date_range}
\title{Get dates of available prices for all tickers in database}
\usage{
get_available_price_date_range(path.database, file.ticker.price.available.db)
}
\arguments{
\item{path.database}{A path where prices are stored.}

\item{file.ticker.price.available.db}{A single character string containing the directory of the database.}
}
\value{
A data frame containing the tickers with dates of available prices.
}
\description{
Get dates of available prices for all tickers in database
}
