% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{get_complete_portfolio_panel}
\alias{get_complete_portfolio_panel}
\title{Get complete portfolio panel}
\usage{
get_complete_portfolio_panel(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\value{
A data frame containing the complete portfolio panel
}
\description{
Get complete portfolio panel
}
