% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_final_tables.R
\name{get_current_portfolio}
\alias{get_current_portfolio}
\title{Get current portfolio investments}
\usage{
get_current_portfolio(path, file.tickers = "isin_ticker.csv",
                      file.transactions = "transaction_fullhistory.csv")
}
\arguments{
\item{path}{A single character string. Directory of your data.}

\item{file.tickers}{A single character string. Name of csv containing ISIN-ticker pairs.}

\item{file.transactions}{A single character string. Name of csv containing all transactions.}
}
\description{
Get current portfolio investments
}
