% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_stock_splits.R
\name{get_current_stock_splits}
\alias{get_current_stock_splits}
\title{Get current stock splits}
\usage{
get_current_stock_splits(user_path)
}
\arguments{
\item{user_path}{A single character string. Directory of user data.}
}
\value{
A data frame containing the current stock splits
}
\description{
Get current stock splits
}
