% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_names.R
\name{get_db_main_path}
\alias{get_db_main_path}
\title{Get database path for main directory}
\usage{
get_db_main_path(db_path = ".")
}
\arguments{
\item{db_path}{A single character string. Path where database is stored.}
}
\description{
Get database path for main directory
}
