% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_target_shares.R
\name{get_default_target_shares}
\alias{get_default_target_shares}
\title{Get default target shares}
\usage{
get_default_target_shares(path)
}
\arguments{
\item{path}{A single character string. Folder where all data are stored.}
}
\value{
A data frame containing a column for \code{isin} (character)
and \code{target_share} (numeric)
}
\description{
Get default target shares
}
