% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_df_with_selected_time_period}
\alias{get_df_with_selected_time_period}
\title{Get data frame for specified time period}
\usage{
get_df_with_selected_time_period(df, nb_period = NULL, period_type = "max")
}
\arguments{
\item{df}{A data frame containing a panel.}

\item{nb_period}{An integer indicating the number of months. Default is
\emph{NULL}. Does not need to be specified for period_type \emph{max} and \emph{ytd}.}

\item{period_type}{A single character string. Default \emph{max}. Possible
values \emph{max}, \emph{ytd}, \emph{weeks} and \emph{months}.}
}
\value{
A data frame containing the original data frame only for the specified time period.
}
\description{
Get data frame for specified time period
}
