% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dividends.R
\name{get_dividends}
\alias{get_dividends}
\title{Get dividend history}
\usage{
get_dividends(path)
}
\arguments{
\item{path}{A single character string. Directory of your data.}
}
\value{
A data frame containing dividend history.
}
\description{
Get dividend history
}
