% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dividends.R
\name{get_dividends_by_year}
\alias{get_dividends_by_year}
\title{Get dividends by year}
\usage{
get_dividends_by_year(path)
}
\arguments{
\item{path}{A single character string. Directory of your data.}
}
\value{
A data frame containing dividends by year in EUR.
}
\description{
Get dividends by year
}
