% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_stats.R
\name{get_dividends_max}
\alias{get_dividends_max}
\title{Get total dividend payments}
\usage{
get_dividends_max(path, file.dividend.history = "dividends_fullhistory.csv")
}
\arguments{
\item{path}{A single character string. Directory where all data are stored.}

\item{file.dividend.history}{A single character string. Name of csv containing full history of dividends.}
}
\description{
Get total dividend payments
}
