% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_names.R
\name{get_file_names}
\alias{get_file_names}
\title{Get file names for PortfolioTracker}
\usage{
get_file_names(path)
}
\arguments{
\item{path}{A single character string. A directory where all data are stored.}
}
\value{
\code{get_file_names} returns a list with file names.
}
\description{
Get file names for PortfolioTracker
}
