% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_investment_irr.R
\name{get_investment_irr}
\alias{get_investment_irr}
\title{Get investment IRR for all tickers}
\usage{
get_investment_irr(path)
}
\arguments{
\item{path}{A single character string. Directory of the data.}
}
\value{
A data frame containing IRRs for all tickers
}
\description{
Get investment IRR for all tickers
}
