% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isins_missing_tickers.R
\name{get_isins_missing_tickers}
\alias{get_isins_missing_tickers}
\title{Get ISINs with missing tickers}
\usage{
get_isins_missing_tickers(path)
}
\arguments{
\item{path}{A single character string. Directory of your data.}
}
\value{
df.isins.missing.tickers A data frame with ISINS for which a ticker was not found.
}
\description{
Get ISINs with missing tickers
}
