% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Get path and file names for PortfolioTracker}
\usage{
get_names(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Get path and file names for PortfolioTracker
}
