% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_names.R
\name{get_path_names}
\alias{get_path_names}
\title{Get path names for PortfolioTracker}
\usage{
get_path_names(path)
}
\arguments{
\item{path}{A single character string. A directory where all data are stored.}
}
\value{
\code{get_path_names} returns a list with path names.
}
\description{
Get path names for PortfolioTracker
}
