% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{get_portfolio_irr}
\alias{get_portfolio_irr}
\title{Get internal rate of return (IRR) of total portfolio}
\usage{
get_portfolio_irr(path, nb_period = NULL, period_type = "max")
}
\arguments{
\item{path}{A single character string. Path where data are stored.}

\item{nb_period}{An integer indicating the number of months. Default
is \emph{NULL}.}

\item{period_type}{A single character string. Default \emph{max}.
Possible values \emph{max}, \emph{weeks} and \emph{months}.}
}
\value{
A numeric for the internal rate of return (IRR) of your portfolio
}
\description{
Get internal rate of return (IRR) of total portfolio
}
