% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portfolio_stats.R
\name{get_portfolio_stats}
\alias{get_portfolio_stats}
\title{Get portfolio statistics}
\usage{
get_portfolio_stats(path)
}
\arguments{
\item{path}{A single character string. Directory of the data.}
}
\value{
A data frame containing basic statistics of the portfolio.
}
\description{
Get portfolio statistics
}
