% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_previous_investments.R
\name{get_previous_investments}
\alias{get_previous_investments}
\title{Get previous investments}
\usage{
get_previous_investments(path)
}
\arguments{
\item{path}{A single character string. Directory of your data.}
}
\value{
A data frame containing current portfolio.
}
\description{
Get previous investments
}
