% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_previous_stock_splits.R
\name{get_previous_stock_splits}
\alias{get_previous_stock_splits}
\title{Get previous stock splits}
\usage{
get_previous_stock_splits(user_path)
}
\arguments{
\item{user_path}{A single character string. Directory of user data.}
}
\value{
A data frame containing the previous stock splits
}
\description{
Get previous stock splits
}
