% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{get_price_panel}
\alias{get_price_panel}
\title{Get full history of prices}
\usage{
get_price_panel(df.transactions, path)
}
\arguments{
\item{df.transactions}{A data.frame containing transaction history.}

\item{path}{A single character string.}
}
\value{
get_price_panel returns
}
\description{
Get full history of prices
}
