% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{get_price_quantity_panel}
\alias{get_price_quantity_panel}
\title{Get panel for the product of prices and quantity}
\usage{
get_price_quantity_panel(ticker, path)
}
\arguments{
\item{ticker}{A single character string containing a ticker symbol.}

\item{path}{A single character string.}
}
\value{
get_price_quantity_panel returns
}
\description{
Get panel for the product of prices and quantity
}
