% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{get_price_quantity_panel2}
\alias{get_price_quantity_panel2}
\title{Get panel for the product of prices and quantity for input ticker}
\usage{
get_price_quantity_panel2(ticker, df_transactions_with_tickers,
                                 path.prices.db)
}
\arguments{
\item{ticker}{A single character string containing a ticker symbol.}

\item{df_transactions_with_tickers}{A data frame containing transactions
and ticker.}

\item{path.prices.db}{A single character string containing the directory of
the price database.}
}
\value{
A data frame with price quantity panel.
}
\description{
Get panel for the product of prices and quantity for input ticker
}
