% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{get_price_quantity_panels}
\alias{get_price_quantity_panels}
\title{Get panels for the product of prices and quantity for all tickers}
\usage{
get_price_quantity_panels(df.transactions, path)
}
\arguments{
\item{df.transactions}{A data.frame containing transaction history.}

\item{path}{A single character string.}
}
\value{
get_price_quantity_panels returns
}
\description{
Get panels for the product of prices and quantity for all tickers
}
