% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{get_prices_from_yahoo}
\alias{get_prices_from_yahoo}
\title{Get price data from Yahoo Finance}
\usage{
get_prices_from_yahoo(ticker, from, to)
}
\arguments{
\item{ticker}{A single character string. Ticker symbol.}

\item{from}{A single character string. Start date.}

\item{to}{A single character string. End date.}
}
\description{
Get price data from Yahoo Finance
}
