% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{get_prices_simple}
\alias{get_prices_simple}
\title{Get prices from the Yahoo Finance API fast and simple}
\usage{
get_prices_simple(ticker, from, to)
}
\arguments{
\item{ticker}{A single character string. Ticker symbol.}

\item{from}{A single character string. Start date.}

\item{to}{A single character string. End date.}
}
\value{
A data frame containing dates, prices (open, high, low, close, adjusted) and volume.
}
\description{
Get prices from the Yahoo Finance API fast and simple
}
