% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{get_quantity_panel}
\alias{get_quantity_panel}
\title{Get full history of quantities}
\usage{
get_quantity_panel(df.transaction.history, path, file.ticker = "isin_ticker.csv")
}
\arguments{
\item{df.transaction.history}{A data.frame containing the full history of transactions.}

\item{path}{A single character string. Directory of your data.}

\item{file.ticker}{A single character string. Name of csv containing ISIN-ticker pairs.}
}
\value{
get_quantity_panel returns
}
\description{
Get full history of quantities
}
