% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{get_returns_all}
\alias{get_returns_all}
\title{Get all returns (daily, monthly, annual) for ticker}
\usage{
get_returns_all(df, ticker)
}
\arguments{
\item{df}{A data frame containing \emph{date} and prices with name \emph{adjusted}.}

\item{ticker}{A single character string containing the ticker.}
}
\value{
A data frame containing daily, monthly and annual returns.
}
\description{
Get all returns (daily, monthly, annual) for ticker
}
