% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{get_roi_varying_periods}
\alias{get_roi_varying_periods}
\title{Get return on investment for specific period}
\usage{
get_roi_varying_periods(df.roi, period)
}
\arguments{
\item{df.roi}{A data.frame containing the complete panel.}

\item{period}{An integer indicating the number of months.}
}
\value{
df.roi.period A data.frame containing return on investment for period.
}
\description{
Get return on investment for specific period
}
