% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ticker.R
\name{get_ticker_from_investing}
\alias{get_ticker_from_investing}
\title{Get ticker from ISIN by crawling investing.com}
\usage{
get_ticker_from_investing(isin, preferred.stock.exchange = "Xetra")
}
\description{
Get ticker from ISIN by crawling investing.com
}
