% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ticker.R
\name{get_ticker_from_xetra}
\alias{get_ticker_from_xetra}
\title{Get ticker based on ISIN by crawling the Xetra website}
\usage{
get_ticker_from_xetra(isin, preferred_exchange = "")
}
\arguments{
\item{isin}{A single character string. ISIN.}

\item{preferred_exchange}{A single character string. Stock exchange (default is empty string)}
}
\description{
Get ticker based on ISIN by crawling the Xetra website
}
