% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tickers_from_db}
\alias{get_tickers_from_db}
\title{Get transactions with tickers}
\usage{
get_tickers_from_db(df_transactions, db_path)
}
\arguments{
\item{df_transactions}{A data frame containing the transaction history.}

\item{db_path}{A single character string containing the directory of the database.}
}
\value{
A list with a data frame containing transactions including the ticker
and a vector of all unique tickers in this data frame.
}
\description{
Get transactions with tickers
}
