% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transactions.R
\name{get_transaction_history}
\alias{get_transaction_history}
\title{Get transaction history}
\usage{
get_transaction_history(path)
}
\arguments{
\item{path}{A single character string. Directory of your data.}
}
\value{
A data frame containing current portfolio.
}
\description{
Get transaction history
}
