% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions_statistics.R
\name{get_transaction_statistics}
\alias{get_transaction_statistics}
\title{get transaction statistics}
\usage{
get_transaction_statistics(df_transactions)
}
\arguments{
\item{df_transactions}{A data frame containing transactions.}
}
\value{
A data frame containing transaction statistics.
}
\description{
get transaction statistics
}
