% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{get_ttwror}
\alias{get_ttwror}
\title{Get true time-weighted rate of return (TTWROR) for portfolio}
\usage{
get_ttwror(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.

 @export}
}
\description{
Get true time-weighted rate of return (TTWROR) for portfolio
}
