% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{get_twr_factors}
\alias{get_twr_factors}
\title{Get TWR factors on daily basis for portfolio}
\usage{
get_twr_factors(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Get the true-weighted return factors on a daily basis for the entire portfolio
}
