% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_names.R
\name{get_user_path_names}
\alias{get_user_path_names}
\title{Get path names for user directory}
\usage{
get_user_path_names(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Get path names for user directory
}
