% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ticker.R
\name{init_isin_ticker}
\alias{init_isin_ticker}
\title{Create ISIN-ticker table as csv if no such file exists}
\usage{
init_isin_ticker(path, file = "isin_ticker.csv")
}
\arguments{
\item{path}{A single character string. Path where data are stored.}

\item{file}{A single character string. Name of ISIN-ticker csv file (Default: isin_ticker.csv)}
}
\description{
Create ISIN-ticker table as csv if no such file exists
}
