% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_portfolio_tracker.R
\name{init_portfolio_tracker}
\alias{init_portfolio_tracker}
\title{Initiate Portfolio Tracker}
\usage{
init_portfolio_tracker(db_path = ".")
}
\arguments{
\item{db_path}{A single character string containing the directory of the database.
Default is the current working directory.}
}
\description{
Initiate Portfolio Tracker (global variables of path names
and file names are loaded, directory with required folders are created)
}
