% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfoliotracker.R
\docType{package}
\name{PortfolioTracker}
\alias{PortfolioTracker}
\title{\code{PortfolioTracker} package}
\description{
Track Your Investment Portfolio
}
\details{
See the README on
\href{https://github.com/lorenzbr/PortfolioTracker#readme}{GitHub}
}
