% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_panels.R
\name{read_panels}
\alias{read_panels}
\title{Read panels}
\usage{
read_panels(path_panels)
}
\arguments{
\item{path_panels}{A single character string. Path of panels.}
}
\value{
A list containing data frames with panels.
}
\description{
Read panels
}
