% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{store_annual_returns}
\alias{store_annual_returns}
\title{Store annual returns for all tickers as csv's}
\usage{
store_annual_returns(path)
}
\arguments{
\item{path}{A single character string. Folder where all data are stored.}
}
\description{
Store annual returns for all tickers as csv's
}
