% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{update_latest_prices}
\alias{update_latest_prices}
\title{Update most recent prices and store in csv files}
\usage{
update_latest_prices(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Update prices for all tickers in the specified path. A new csv is created
for the updated prices. Better use \code{\link{append_latest_prices}} which is
more efficient.
}
