% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{write_annualized_returns}
\alias{write_annualized_returns}
\title{Write annualized returns to a csv file}
\usage{
write_annualized_returns(user_path, db_path)
}
\arguments{
\item{user_path}{A single character string containing the directory of the user.}

\item{db_path}{A single character string containing the directory of the database.}
}
\description{
Write annualized returns to a csv file
}
