% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_complete_panel}
\alias{write_complete_panel}
\title{Write complete panel to a csv file}
\usage{
write_complete_panel(ticker, user_path)
}
\arguments{
\item{ticker}{A single character string containing the ticker.}

\item{user_path}{A single character string. Path where data are stored.}
}
\description{
Write complete panel to a csv file
}
