% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_complete_panels}
\alias{write_complete_panels}
\title{Write complete panels to a csv file}
\usage{
write_complete_panels(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Write complete panels to a csv file
}
