% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{write_cum_investment_returns_daily}
\alias{write_cum_investment_returns_daily}
\title{Write cumulative daily investment returns to a csv file}
\usage{
write_cum_investment_returns_daily(path)
}
\arguments{
\item{path}{A single character string. Directory where all data are stored.}
}
\description{
Write cumulative daily investment returns to a csv file
}
