% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_portfolio.R
\name{write_current_portfolio}
\alias{write_current_portfolio}
\title{Write current portfolio investments to a csv file}
\usage{
write_current_portfolio(user_path, db_path)
}
\arguments{
\item{user_path}{A single character string containing the directory of the user.}

\item{db_path}{A single character string containing the directory of the database.}
}
\description{
Write current portfolio investments to a csv file
}
