% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_portfolio.R
\name{write_default_target_shares}
\alias{write_default_target_shares}
\title{Write default target shares to csv file}
\usage{
write_default_target_shares(df.target.shares, path)
}
\arguments{
\item{df.target.shares}{A data frame containing a column for \code{isin} (character) and \code{target_share} (numeric)}

\item{path}{A single character string. Folder where all data are stored.}
}
\description{
Write default target shares to csv file
}
