% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_dividends.R
\name{write_dividend_by_month}
\alias{write_dividend_by_month}
\title{Write dividend payments by month to a csv file}
\usage{
write_dividend_by_month(path,
           file.dividend.history = "dividends_fullhistory.csv")
}
\arguments{
\item{path}{A single character string. Directory where all data are stored.}

\item{file.dividend.history}{A single character string. Name of csv containing full history of dividends.}
}
\description{
Write dividend payments by month to a csv file
}
