% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_dividends.R
\name{write_dividend_history}
\alias{write_dividend_history}
\title{Write history of dividends in a csv file}
\usage{
write_dividend_history(df_transactions, path)
}
\arguments{
\item{df_transactions}{A data frame containing the history of transactions.}

\item{path}{A single character string. Path where data are stored.}
}
\description{
Write history of dividends in a csv file
}
