% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{write_investment_irr_all}
\alias{write_investment_irr_all}
\title{Write IRRs for all tickers to csv files}
\usage{
write_investment_irr_all(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Write IRRs for all tickers to csv files
}
