% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_investment_value_panel}
\alias{write_investment_value_panel}
\title{Write investment value panel for given ticker to csv file}
\usage{
write_investment_value_panel(ticker, path)
}
\arguments{
\item{ticker}{A single character string containing the ticker symbol.}

\item{path}{A single character string. Path where data are stored.}
}
\description{
The function writes the investment value for a given \emph{ticker}
and day as a csv file to \emph{path}. The investment value is defined as the sum
of the current value of the investment including current value of a sale and
dividends excluding current costs from purchasing the investment.
See \code{\link{write_investment_value_panels}} if you want to apply this function to all
tickers in your table of transactions.
}
