% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_investment_value_panels}
\alias{write_investment_value_panels}
\title{Write investment value panels for all tickers to csv files}
\usage{
write_investment_value_panels(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
This functions writes investment value panels for all tickers
to separate csv files. See \code{\link{write_investment_value_panel}} for further information.
}
