% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{write_portfolio_return}
\alias{write_portfolio_return}
\title{Write portfolio return to a csv file}
\usage{
write_portfolio_return(path)
}
\arguments{
\item{path}{A single character string. Directory where all data are stored.}
}
\description{
Write portfolio return to a csv file
}
