% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_stats.R
\name{write_portfolio_stats}
\alias{write_portfolio_stats}
\title{Write portfolio statistics}
\usage{
write_portfolio_stats(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Write portfolio statistics
}
