% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_price_panels}
\alias{write_price_panels}
\title{Write full history of prices for all tickers as csv}
\usage{
write_price_panels(df.transactions, path)
}
\arguments{
\item{df.transactions}{A data.frame containing transaction history.}

\item{path}{A single character string containing the directory of the project.}
}
\description{
Write full history of prices for all tickers as csv
}
