% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_price_quantity_panel}
\alias{write_price_quantity_panel}
\title{Write panel for the product of prices and quantity for input ticker as csv}
\usage{
write_price_quantity_panel(ticker, path)
}
\arguments{
\item{ticker}{A single character string containing a ticker symbol.}

\item{path}{A single character string containing the directory of the project.}
}
\description{
Write panel for the product of prices and quantity for input ticker as csv
}
