% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_price_quantity_panels_db}
\alias{write_price_quantity_panels_db}
\title{Write panels for the product of prices and quantity for all tickers as csv}
\usage{
write_price_quantity_panels_db(df.transactions, user_path, db_path)
}
\arguments{
\item{df.transactions}{A data frame containing transaction history.}

\item{user_path}{A single character string containing the directory of the user.}

\item{db_path}{A single character string containing the directory of the database.}
}
\description{
Write panels for the product of prices and quantity for all tickers as csv
}
