% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_quantity_panel}
\alias{write_quantity_panel}
\title{Write quantity panel for ticker}
\usage{
write_quantity_panel(ticker, df_transactions_with_tickers,
                            path.quantity.panel)
}
\arguments{
\item{ticker}{A single character string containing the ticker symbol.}

\item{df_transactions_with_tickers}{A data frame containing transactions
and ticker.}

\item{path.quantity.panel}{A single character string containing the folder
of quantity panels.}
}
\description{
Write quantity panel for ticker
}
