% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{write_returns}
\alias{write_returns}
\title{Write returns for all tickers to csv files}
\usage{
write_returns(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Write returns for all tickers to csv files
}
