% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{write_roi_by_period}
\alias{write_roi_by_period}
\title{Write cumulative daily investment returns to a csv file}
\usage{
write_roi_by_period(ticker, path)
}
\arguments{
\item{ticker}{A single character string containing the ticker.}

\item{path}{A single character string. Path where data are stored.}
}
\description{
Write cumulative daily investment returns to a csv file
}
