% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{write_twr_factors}
\alias{write_twr_factors}
\title{Write TWR factors on daily basis for portfolio}
\usage{
write_twr_factors(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Write the true-weighted return factors on a daily basis for the entire portfolio
}
