% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfoliotracker_dir.R
\name{create_portfoliotracker_dir}
\alias{create_portfoliotracker_dir}
\title{Create directory for portfoliotracker}
\usage{
create_portfoliotracker_dir(path)
}
\arguments{
\item{path}{A single character string. Folder where all data are stored.}
}
\description{
Create directory for portfoliotracker
}
