% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{get_annual_returns}
\alias{get_annual_returns}
\title{Get annual returns}
\usage{
get_annual_returns(df)
}
\arguments{
\item{df}{A data frame containing date, adjusted prices and ticker.}
}
\value{
df.annualreturns A data frame containing years and annual returns.
}
\description{
Get annual returns
}
