% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returns.R
\name{get_annualized_returns}
\alias{get_annualized_returns}
\title{Get annualized returns}
\usage{
get_annualized_returns(path)
}
\arguments{
\item{path}{A single character string. Directory of the data.}
}
\value{
A data frame containing formatted table with annualized returns
for all tickers.
}
\description{
Get annualized returns
}
