% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_portfolio.R
\name{get_current_portfolio}
\alias{get_current_portfolio}
\title{Get current portfolio}
\usage{
get_current_portfolio(path)
}
\arguments{
\item{path}{A single character string. Directory of your data.}
}
\value{
A data frame containing the current investments.
}
\description{
Get current portfolio
}
