% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isins_missing_tickers.R
\name{get_isins_missing_tickers}
\alias{get_isins_missing_tickers}
\title{Get ISINs with missing tickers}
\usage{
get_isins_missing_tickers(user_path, db_path)
}
\arguments{
\item{user_path}{A single character string containing the directory of the user.}

\item{db_path}{A single character string containing the directory of the database.}
}
\value{
A data frame with ISIN-ticker pairs for which a ticker was not found.
}
\description{
Get ISINs with missing tickers
}
