% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{get_prices_from_yahoo}
\alias{get_prices_from_yahoo}
\title{Get prices for a given ticker from the Yahoo Finance API}
\usage{
get_prices_from_yahoo(ticker, from, to, preferred_exchange = "Xetra",
                             stock_exchanges = c(".DE", ".F", ".SG", ".MU", ".DU"),
                             method = "simple", user_specific_exchange = TRUE)
}
\arguments{
\item{ticker}{A single character string. Ticker symbol.}

\item{from}{A single character string. Start date.}

\item{to}{A single character string. End date.}

\item{preferred_exchange}{A single character string. Stock exchange
(default is \emph{Xetra})}

\item{stock_exchanges}{A vector of single character strings. Possible stock
exchanges to get prices from.}

\item{method}{A single character string (default: \emph{simple}). The method
how to get prices from Yahoo. Use the R package \emph{quantmod} or a \emph{simple}
and fast solution.}

\item{user_specific_exchange}{Logical indicating whether prices are taken from
exchange that is preferred by the user (i.e., stored in the user's profile).
Default is \emph{TRUE}.}
}
\value{
A data frame containing dates, prices (open, high, low, close, adjusted) and volume.
}
\description{
Get prices for a given ticker from the Yahoo Finance API. Wrapper around
\code{\link{get_ticker_prices}} with further arguments. E.g., select the stock exchange.
}
