% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{get_quantity_panel}
\alias{get_quantity_panel}
\title{Get quantity panel for ticker}
\usage{
get_quantity_panel(ticker, df_transactions_with_tickers)
}
\arguments{
\item{ticker}{A single character string containing the ticker symbol.}

\item{df_transactions_with_tickers}{A data frame containing transactions
and ticker.}
}
\value{
A data frame with a quantity panel.
}
\description{
Get quantity panel for ticker
}
