% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ticker.R
\name{get_ticker_from_investing}
\alias{get_ticker_from_investing}
\title{Get ticker based on ISIN by crawling investing.com}
\usage{
get_ticker_from_investing(isin, preferred.stock.exchange = "")
}
\arguments{
\item{isin}{A single character string. ISIN.}

\item{preferred.stock.exchange}{A single character string. Stock exchange (e.g., Xetra)}
}
\description{
Get ticker based on ISIN by crawling investing.com
}
