% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ticker.R
\name{get_ticker_from_investing}
\alias{get_ticker_from_investing}
\title{Get ticker based on ISIN by crawling investing.com}
\usage{
get_ticker_from_investing(isin, preferred_exchange = "")
}
\arguments{
\item{isin}{A single character string. ISIN.}

\item{preferred_exchange}{A single character string. Stock
exchange (e.g., Xetra)}
}
\value{
A single character string containing the ticker for the given ISIN.
}
\description{
Get ticker based on ISIN by crawling investing.com
}
