% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_user_names}
\alias{get_user_names}
\title{Get user-specific path and file names}
\usage{
get_user_names(path)
}
\arguments{
\item{path}{A single character string. Path where user-specific data are stored.}
}
\description{
Get user-specific path and file names
}
