% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_names.R
\name{get_user_path_names}
\alias{get_user_path_names}
\title{Get path names for user directory}
\usage{
get_user_path_names(path, portfolio_name = "portfolio_1")
}
\arguments{
\item{path}{A single character string. Path where data are stored.}

\item{portfolio_name}{A single character string with the name of the
portfolio. All portfolio-specific data for the focal user are stored in
a folder with this name. Default is \emph{portfolio_1}.}
}
\description{
Get path names for user directory
}
