% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_portfolio.R
\name{rebalance_portfolio}
\alias{rebalance_portfolio}
\title{Produces a suggestion to rebalance your portfolio}
\usage{
rebalance_portfolio(df, df_target_shares, money_to_invest, step_size = 500,
                           exclude_from_rebalancing = NA)
}
\arguments{
\item{df}{A data frame containing at least a column for \code{isin}
(character) and \code{value} (numeric)}

\item{df_target_shares}{A data frame containing a column for \code{isin}
(character) and \code{target_share} (numeric)}

\item{money_to_invest}{A single numeric. Amount to invest for rebalancing.}

\item{step_size}{A single numeric. Step size (default is 500).}

\item{exclude_from_rebalancing}{A single character or vector of strings.
Should contain ISINs (default is NA).}
}
\description{
Produces a suggestion to rebalance your portfolio
}
