% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_portfolio.R
\name{rebalance_portfolio}
\alias{rebalance_portfolio}
\title{Produces a suggestion to rebalance your portfolio}
\usage{
rebalance_portfolio(
  df,
  df.target.shares,
  exclude.from.rebalancing,
  money.to.invest,
  step.size = 500
)
}
\description{
Produces a suggestion to rebalance your portfolio
}
