% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{update_latest_prices}
\alias{update_latest_prices}
\title{Update most recent prices and store as csv}
\usage{
update_latest_prices(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Update most recent prices and store as csv
}
