% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{update_prices_based_on_transactions}
\alias{update_prices_based_on_transactions}
\title{Update and store prices as csv based on new financial transactions}
\usage{
update_prices_based_on_transactions(df.transactions, path, external.search = TRUE)
}
\arguments{
\item{df.transactions}{A data frame. Results from [BankStatementParser::get_transactions()]}

\item{path}{A single character string. Path where data are stored.}

\item{external.search}{Logical if TRUE, the function searches external sources to find the ticker.}
}
\description{
Update and store prices as csv based on new financial transactions
}
