% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{update_prices_based_on_transactions}
\alias{update_prices_based_on_transactions}
\title{Update and store prices as csv based on new financial transactions}
\usage{
update_prices_based_on_transactions(
  path.prices,
  path.input.transactions,
  path.output.transactions
)
}
\description{
Update and store prices as csv based on new financial transactions
}
