% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_all_value_panels}
\alias{write_all_value_panels}
\title{Write all value panels for purchase, dividend and sales transactions for all tickers}
\usage{
write_all_value_panels(df.transaction.history, path)
}
\arguments{
\item{df.transaction.history}{A data frame containing the full history of transactions.}

\item{path}{A single character string containing the directory of the project.}
}
\description{
Write all value panels for purchase, dividend and sales transactions for all tickers
}
