% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{write_annual_returns}
\alias{write_annual_returns}
\title{Write annual returns for all tickers as csv's}
\usage{
write_annual_returns(path)
}
\arguments{
\item{path}{A single character string. Directory where all data are stored.}
}
\description{
Write annual returns for all tickers as csv's
}
