% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{write_annualized_returns}
\alias{write_annualized_returns}
\title{Write annualized returns to a csv file}
\usage{
write_annualized_returns(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Write annualized returns to a csv file
}
