% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_complete_panels}
\alias{write_complete_panels}
\title{Write complete panels to a csv file}
\usage{
write_complete_panels(user_path, db_path)
}
\arguments{
\item{user_path}{A single character string containing the directory of the user.}

\item{db_path}{A single character string containing the directory of the database.}
}
\description{
Write complete panels to a csv file
}
