% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_portfolio.R
\name{write_current_portfolio}
\alias{write_current_portfolio}
\title{Write current portfolio investments to a csv file}
\usage{
write_current_portfolio(path)
}
\arguments{
\item{path}{A single character string. Directory of your data.}
}
\description{
Write current portfolio investments to a csv file
}
