% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_dividends.R
\name{write_dividend_by_month}
\alias{write_dividend_by_month}
\title{Write dividend payments by month to a csv file}
\usage{
write_dividend_by_month(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Write dividend payments by month to a csv file
}
