% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_dividends.R
\name{write_dividend_by_yr}
\alias{write_dividend_by_yr}
\title{Write dividend payments by year to a csv file}
\usage{
write_dividend_by_yr(path)
}
\arguments{
\item{path}{A single character string. Directory where all data are stored.}
}
\description{
Write dividend payments by year to a csv file
}
