% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_investment_value_panel}
\alias{write_investment_value_panel}
\title{Write investment value panel for given ticker}
\usage{
write_investment_value_panel(ticker, path)
}
\arguments{
\item{ticker}{A single character string containing the ticker symbol.}

\item{path}{A single character string. Path where data are stored.}
}
\description{
Write investment value panel for given ticker
}
