% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_investment_value_panels}
\alias{write_investment_value_panels}
\title{Write all investment value panels to a csv file}
\usage{
write_investment_value_panels(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Write all investment value panels to a csv file
}
