% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous_investments.R
\name{write_previous_investments}
\alias{write_previous_investments}
\title{Write previous portfolio investments to a csv file}
\usage{
write_previous_investments(path)
}
\arguments{
\item{path}{A single character string. Directory of your data.}
}
\description{
Write previous portfolio investments to a csv file
}
