% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous_investments.R
\name{write_previous_investments}
\alias{write_previous_investments}
\title{Write previous portfolio investments to a csv file}
\usage{
write_previous_investments(user_path, db_path)
}
\arguments{
\item{user_path}{A single character string containing the directory of the user.}

\item{db_path}{A single character string containing the directory of the database.}
}
\description{
Write previous portfolio investments to a csv file
}
