% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_price_panels}
\alias{write_price_panels}
\title{Write price panels for all tickers to csv files}
\usage{
write_price_panels(df_transactions, path)
}
\arguments{
\item{df_transactions}{A data frame containing transaction history.}

\item{path}{A single character string containing the directory of the user.}
}
\description{
Write price panels for all tickers to csv files
}
