% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_price_quantity_panel_db}
\alias{write_price_quantity_panel_db}
\title{Write panel for the product of prices and quantity for input ticker as csv}
\usage{
write_price_quantity_panel_db(ticker, df_transactions_with_tickers,
                                     user_path, db_path)
}
\arguments{
\item{ticker}{A single character string containing a ticker symbol.}

\item{df_transactions_with_tickers}{A data frame containing transactions
and ticker.}

\item{user_path}{A single character string containing the directory of the user.}

\item{db_path}{A single character string containing the directory of the database.}
}
\description{
Write panel for the product of prices and quantity for input ticker as csv
}
