% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_returns.R
\name{write_roi_by_period_all}
\alias{write_roi_by_period_all}
\title{Write cumulative daily investment returns for all tickers to a csv file}
\usage{
write_roi_by_period_all(path)
}
\arguments{
\item{path}{A single character string. Path where data are stored.}
}
\description{
Write cumulative daily investment returns for all tickers to a csv file
}
