% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{write_value_panel}
\alias{write_value_panel}
\title{Write value panel for purchase, dividend and sales transactions for given ticker}
\usage{
write_value_panel(transaction_type, ticker,
                         df_transactions, user_path)
}
\arguments{
\item{transaction_type}{A single character string containing the transaction
type (e.g., \emph{Purchase}, \emph{Sale} or \emph{Dividend})}

\item{ticker}{A single character string containing the ticker symbol.}

\item{df_transactions}{A data frame containing the full history of transactions.}

\item{user_path}{A single character string containing the directory of the user.}
}
\description{
Write value panel for purchase, dividend and sales transactions for given ticker
}
